moved recommender out of nextjs
