+++ recommender/recommender_codes/user/find_similar_movies_coll.py
