+++ src/app/page.tsx
@@ -5,8 +5,8 @@ import { MovieList } from "@/components/movieList";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/authOptions";
 import GenreSelector from "@/components/genreSelector";
-import { getUserRecommendations } from "@/recommender/getUserRecommendations";
-import { isRecommendations } from "@/recommender/isRecommendations";
+import { getUserRecommendations } from "@/services/getUserRecommendations";
+import { isRecommendations } from "@/services/isRecommendations";
 
 export default async function Home({ searchParams }: { searchParams?: { genre: string } }) {
   const selectedGenre = searchParams?.genre;
