+++ src/services/getUserRecommendations.tsx
@@ -7,7 +7,7 @@ export const getUserRecommendations = async (selectedGenre: string | undefined, 
   for (const rating of ratingsFavourites.ratings) {
     ratingsObject[rating.movieId] = rating.rating;
   }
-  const response = await fetch("http://localhost:5000/recommender/user", {
+  const response = await fetch("http://recommender:5000/recommender/user", {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
